Support bitbucket tags on git pull option
I found an error when someone push a tag to bitbucket because the path where we execute the pull is incorrect formed like it would be a regular branch.
I added an if checking the type of change that comes in the json form bitbucket, if the type is "tag" I add to the "branch_name" "tags/". This will be used on the pull_repo method to form the reference correctly.